Alteração dos tipos da variável j, e dos vectors prefixo e posicoes, assim como os retornos das funcoes buscarKMP e construir prefixo :
troca:
int -> size_t

Motivo:
Ao compilar, a comparação de tipos sem sinal (size_t) e com sinal (int) causa warnings.

em:
for (size_t i = 0; i < texto.length(); i++) {
o i que era int era comparado com o metodo .length que retorna um size_t

tambem acontecia para o j

+++ kmp.cpp
@@ -4,9 +4,9 @@
 using namespace std;
 
 // Função que constroi vetor de prefixos
-std::vector<int> construirPrefixo(const std::string& padrao) {
+std::vector<size_t> construirPrefixo(const std::string& padrao) {
     int m = padrao.length(), j = 0;
-    vector<int> prefixo(m);
+    vector<size_t> prefixo(m);
     
     for (int i = 1; i < m; i++) {
         while (j > 0 && padrao[i] != padrao[j])
@@ -20,12 +20,12 @@ std::vector<int> construirPrefixo(const std::string& padrao) {
 }
 
 // Função KMP
-std::vector<int> buscarKMP(const std::string& texto, const std::string& padrao) {
-    vector<int> posicoes;
-    vector<int> prefixo = construirPrefixo(padrao);
-    int j = 0;
+std::vector<size_t> buscarKMP(const std::string& texto, const std::string& padrao) {
+    vector<size_t> posicoes;
+    vector<size_t> prefixo = construirPrefixo(padrao);
+    size_t j = 0;
 
-    for (int i = 0; i < texto.length(); i++) {
+    for (size_t i = 0; i < texto.length(); i++) {
         while (j > 0 && texto[i] != padrao[j])
             j = prefixo[j - 1];
         if (texto[i] == padrao[j])
